Refactor server actions to use JWT authentication
